requirements.txt
asgiref==3.9.1
django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.7.0
django-dotenv==1.4.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
pillow==11.3.0
pyjwt==2.10.1
pytz==2025.2
python-dotenv==1.1.1